# server/swagger.yaml
openapi: 3.0.0
info:
  title: BitcoYemen API
  contact: {}
  version: "V1.2"
servers:
  - url: https://bitco-ye.onrender.com/api/v1
    variables: {}
  - url: https://bitco-ye.herokuapp.com/api/v1
    variables: {}
  - url: http://localhost:5000/api/v1
    variables: {}
paths:
...
  /users/updateUser:
    patch:
      tags:
        - User
      summary: Update User
      operationId: UpdateUser
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - address
                - phoneNumber
              type: object
              properties:
                address:
                  type: string
                  example: asda
                phoneNumber:
                  type: integer
                  format: int32
                  example: 7896
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /users/updateUserPassword:
    patch:
      tags:
        - User
      summary: Update User Password
      operationId: UpdateUserPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateUserPasswordRequest"
                - example:
                    oldPassword: "123456"
                    newPassword: "1234567"
            example:
              oldPassword: "123456"
              newPassword: "1234567"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /users/updateUser/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    post:
      tags:
        - User
      summary: Update User Admin
      operationId: UpdateUserAdmin
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - address
                - phoneNumber
                - name
                - role
              type: object
              properties:
                address:
                  type: string
                  example: asdaa
                phoneNumber:
                  type: integer
                  format: int32
                  example: 7896
                name:
                  type: string
                  example: Omer
                role:
                  type: string
                  example: user
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /users/updateUserPassword/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    post:
      tags:
        - User
      summary: Update User Password Admin
      operationId: UpdateUserPasswordAdmin
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateUserPasswordAdminRequest"
                - example:
                    newPassword: "123456"
            example:
              newPassword: "123456"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /users/activate/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    post:
      tags:
        - User
      summary: Activate User
      operationId: ActivateUser
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ActivateUserRequest"
                - example:
                    isActive: true
            example:
              isActive: true
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /transaction:
    post:
      tags:
        - Transaction
      summary: Create Transaction
      operationId: CreateTransaction
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateTransactionRequest"
                - example:
                    from: 62e4c4a5bb77243116e4c386
                    to: 62e4c4d1bb77243116e4c389
                    amount: "181"
            example:
              from: 62e4c4a5bb77243116e4c386
              to: 62e4c4d1bb77243116e4c389
              amount: "181"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Transaction
      summary: User Transactions
      operationId: UserTransactions
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /transaction/status/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The transaction ID
    post:
      tags:
        - Transaction
      summary: Change Transaction Status
      operationId: ChangeTransactionStatus
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ChangeTransactionStatusRequest"
                - example:
                    statusCode: 5
            example:
              statusCode: 5
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /recommendation:
    post:
      tags:
        - Recommendation
      summary: Create Recommendation
      operationId: CreateRecommendation
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateRecommendationRequest"
                - example:
                    name: USD/EUR
                    price: "1.258"
                    tp1: asdf ';lk
                    tp2: asdf ';lk
                    tp3: asdf ';lk
                    sl: asdf ';lk
                    status: sell
            example:
              name: USD/EUR
              price: "1.258"
              tp1: asdf ';lk
              tp2: asdf ';lk
              tp3: asdf ';lk
              sl: asdf ';lk
              status: sell
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Recommendation
      summary: All Recommendation
      operationId: AllRecommendation
      parameters:
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: limit
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /recommendation/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The recommendation ID
    post:
      tags:
        - Recommendation
      summary: Update Recommendation
      operationId: UpdateRecommendation
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateRecommendationRequest"
                - example:
                    name: USD/EURR
                    price: "1.258"
                    tp1: asdf ';lk
                    tp2: asdf ';lk
                    tp3: asdf ';lk
                    sl: asdf ';lk
                    status: sell
            example:
              name: USD/EURR
              price: "1.258"
              tp1: asdf ';lk
              tp2: asdf ';lk
              tp3: asdf ';lk
              sl: asdf ';lk
              status: sell
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Recommendation
      summary: Recommendation
      operationId: Recommendation
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
    delete:
      tags:
        - Recommendation
      summary: Recommendation1
      operationId: Recommendation1
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
  /setting/available:
    get:
      tags:
        - Settings
      summary: Service Availability
      operationId: ServiceAvailability
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    post:
      tags:
        - Settings
      summary: Set Service Availability
      operationId: SetServiceAvailability
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SetServiceAvailabilityRequest"
                - example:
                    available: false
            example:
              available: false
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    UpdateUserPasswordRequest:
      title: UpdateUserPasswordRequest
      required:
        - oldPassword
        - newPassword
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      example:
        oldPassword: "123456"
        newPassword: "1234567"
    UpdateUserPasswordAdminRequest:
      title: UpdateUserPasswordAdminRequest
      required:
        - newPassword
      type: object
      properties:
        newPassword:
          type: string
      example:
        newPassword: "123456"
    ActivateUserRequest:
      title: ActivateUserRequest
      required:
        - isActive
      type: object
      properties:
        isActive:
          type: boolean
      example:
        isActive: true
...
    CreateTransactionRequest:
      title: CreateTransactionRequest
      required:
        - from
        - to
        - amount
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: string
      example:
        from: 62e4c4a5bb77243116e4c386
        to: 62e4c4d1bb77243116e4c389
        amount: "181"
    ChangeTransactionStatusRequest:
      title: ChangeTransactionStatusRequest
      required:
        - statusCode
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
      example:
        statusCode: 5
...
    CreateRecommendationRequest:
      title: CreateRecommendationRequest
      required:
        - name
        - price
        - tp1
        - tp2
        - tp3
        - sl
        - status
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        tp1:
          type: string
        tp2:
          type: string
        tp3:
          type: string
        sl:
          type: string
        status:
          type: string
      example:
        name: USD/EUR
        price: "1.258"
        tp1: asdf ';lk
        tp2: asdf ';lk
        tp3: asdf ';lk
        sl: asdf ';lk
        status: sell
    UpdateRecommendationRequest:
      title: UpdateRecommendationRequest
      required:
        - name
        - price
        - tp1
        - tp2
        - tp3
        - sl
        - status
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        tp1:
          type: string
        tp2:
          type: string
        tp3:
          type: string
        sl:
          type: string
        status:
          type: string
      example:
        name: USD/EURR
        price: "1.258"
        tp1: asdf ';lk
        tp2: asdf ';lk
        tp3: asdf ';lk
        sl: asdf ';lk
        status: sell
...
    SetServiceAvailabilityRequest:
      title: SetServiceAvailabilityRequest
      required:
        - available
      type: object
      properties:
        available:
          type: boolean
      example:
        available: false
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...